requirements.txt
pandas
PyYAML
pydantic
pydantic-settings
pandera
colorlog
networkx
neo4j
pyvis
sentence-transformers
scikit-learn
chromadb
thefuzz
sentence-transformers
async-lru
pybreaker
colorlog
google-generativeai
spacy
en_core_web_sm
pytest-asyncio-cooperative
pytest-benchmark